% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_valid_subset_input.R
\name{check_for_valid_subset_input}
\alias{check_for_valid_subset_input}
\title{Internal function to check for valid subset input}
\usage{
check_for_valid_subset_input(period, input)
}
\arguments{
\item{period}{Period corresponding to the input to be checked}

\item{input}{Input to be checked}
}
\description{
Many tweetexploR functions allow the user to subset the data
via the \code{to} and \code{from} parameters. Depending on the time period for the
chart they are creating with one of the tweetexploR functions (e.g.,
\code{num_tweets_by_timeperiod()}), a specific date/time format is required, as
follows:

For hourly charts, provide a character string in the format
\code{"\%Y-\%m-\%d \%H:\%M:\%S"}. For example \code{"2022-06-20 06:00:00"}

For daily charts, provide a character string in the format
\code{"\%Y-\%m-\%d"}. For example \code{"2022-06-20"}

For monthly charts, provide a character string in the format
\code{"\%Y-\%m"}. For example \code{"2022-06"}
}
