% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_valid_period.R
\name{check_if_period_is_valid}
\alias{check_if_period_is_valid}
\title{Internal function to check if \code{period} is valid}
\usage{
check_if_period_is_valid(period)
}
\arguments{
\item{period}{Time period to be checked}
}
\description{
If \code{period} is not valid, the program will stop and give a
message. If \code{period} is valid, the program will continue.
}
