% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_ggplot_theme.R
\name{configure_ggplot_theme}
\alias{configure_ggplot_theme}
\title{Internal function to configure ggplot2 theme}
\usage{
configure_ggplot_theme()
}
\value{
ggplot2 theme
}
\description{
Internal function to configure ggplot2 theme
}
