% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_y_axis.R
\name{configure_y_axis}
\alias{configure_y_axis}
\title{Internal function to configure ggplot2 y axis}
\usage{
configure_y_axis()
}
\value{
ggplot2 y axis
}
\description{
Internal function to configure ggplot2 y axis
}
