% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_sqlite_db.R
\name{connect_to_sqlite_db}
\alias{connect_to_sqlite_db}
\title{Connect to tidy-tweet SQLite database}
\usage{
connect_to_sqlite_db(sqlite_file)
}
\arguments{
\item{sqlite_file}{Path to SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The path should end with \code{.db}}
}
\value{
Class SQLiteConnection object (more information is available
\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{here}).
}
\description{
Connect to SQLite database file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables.

It is recommended to create a variable for the \code{SQLiteConnection} created
by \code{connect_to_sqlite_db()} so that you can use it for other tweetexploR
functions.
}
\examples{
\dontrun{

sqlite_con <- connect_to_sqlite_db("my_database.db")
sqlite_con <- connect_to_sqlite_db("my_subfolder\\\\my_database.db")
}

}
