% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engagement_summary.R
\name{engagement_summary}
\alias{engagement_summary}
\title{Generate a summary of engagement metrics for all tweets}
\usage{
engagement_summary(sqlite_con)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}
}
\value{
data frame containing a summary of engagement metrics.
}
\description{
Generate a summary of engagement metrics for all tweets. The
summary will be in the form of a data frame, and will include the
following Twitter metrics: likes, quotes, replies, and retweets. A total of
all Twitter metrics will also be included. The summary will be sorted in
descending order by total, and will include the tweet text, tweet ID,
author ID, and author username. Crucially, there are two date/time columns
returned (as POSIXct in UTC time): created_at and retrieved_at. This allows
the user to normalise engagement metrics against the difference between
created_at and retrieved_at (since more time between a tweet being created
and retrieved means more opportunity for the tweet to accumulate
engagement).
}
\examples{
\dontrun{

engagement_summary(sqlite_con)

engagement_summary <- engagement_summary(sqlite_con)

}

}
