% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_tweets_by_timeperiod.R
\name{num_tweets_by_timeperiod}
\alias{num_tweets_by_timeperiod}
\title{Plot number of tweets per hour, day, or month}
\usage{
num_tweets_by_timeperiod(sqlite_con, period, from, to)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}

\item{period}{can be "hour", "day", or "month". TODO: Default value?}

\item{from}{TODO: Should be optional. Default should be no filter.}

\item{to}{TODO: Should be optional. Default should be no filter.}
}
\value{
ggplot2 plot.
}
\description{
Create a ggplot2 chart of the number of tweets per
hour, day or month.
}
\examples{
\dontrun{

num_tweets_by_timeperiod(sqlite_con, period = "day", from = "2022-06-14",
  to = "2022-06-20")
my_plot <- num_tweets_by_timeperiod(sqlite_con, period = "hour",
  from = "2022-06-20 06:00:00", to = "2022-06-21 06:00:00")
}

}
