% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_tweets_by_username.R
\name{num_tweets_by_username}
\alias{num_tweets_by_username}
\title{Plot number of tweets for top n usernames}
\usage{
num_tweets_by_username(
  sqlite_con,
  n = 10,
  return_data = FALSE,
  exclude_RT = FALSE,
  ...
)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}

\item{n}{Number of usernames to be plotted. Note, ties will be included.
Default value is 10. Recommended value to ensure clear visualisation is
<= 30.}

\item{return_data}{Should the data underlying the chart be returned?
The default is \code{FALSE}. If \code{return_data = TRUE}, the data can be accessed
in the second element, \code{data}, of the returned list.}

\item{exclude_RT}{Should retweets be excluded from the calculations?
The default is \code{FALSE}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}.}
}
\value{
ggplot2 plot. If \code{return_data = TRUE}, returns a named list with the
first element, \code{chart}, being a ggplot2 plot, and the second element,
\code{data}, being the underlying data for the ggplot2 plot in the form of a
data frame.
}
\description{
Create a ggplot2 bar chart of the number of tweets per username
for the top \code{n} usernames (ties for \code{n}th position will be included).
}
\examples{
\dontrun{

num_tweets_by_username(sqlite_con, n = 12)

my_plot <- num_tweets_by_username(sqlite_con, 20)

num_tweets_by_username(sqlite_con, fill = "blue")

results <- num_tweets_by_username(sqlite_con, return_data = TRUE)

results <- num_tweets_by_username(sqlite_con, exclude_RT = TRUE)

}

}
