% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_missing_from_input.R
\name{substitute_missing_from_input}
\alias{substitute_missing_from_input}
\title{Internal function to substitute missing value of \code{from} (optional parameter)
to a very old date in the appropriate format}
\usage{
substitute_missing_from_input(period)
}
\arguments{
\item{period}{Time period for which the tweet frequency will be calculated.
Accepted values are \code{"hour"}, \code{"day"}, or \code{"month"}. Defaults to \code{"day"}.}
}
\value{
from Optional parameter to subset the data to tweets that were
created on or before the specified date and/or time in UTC.
For hourly charts, provide a character string in the format
\code{"\%Y-\%m-\%d \%H:\%M:\%S"}. For example \code{"2022-06-20 06:00:00"}

For daily charts, provide a character string in the format
\code{"\%Y-\%m-\%d"}. For example \code{"2022-06-20"}

For monthly charts, provide a character string in the format
\code{"\%Y-\%m"}. For example \code{"2022-06"}
}
\description{
Internal function to substitute missing value of \code{from} (optional parameter)
to a very old date in the appropriate format
}
