% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_hashtags.R
\name{top_n_hashtags}
\alias{top_n_hashtags}
\title{Plot top n hashtags}
\usage{
top_n_hashtags(sqlite_con, n = 10)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}

\item{n}{Number of hashtags to be plotted. Note, ties will be included.
Default value is 10.}
}
\value{
ggplot2 plot.
}
\description{
Create a ggplot2 bar chart of the number of times the top \code{n}
hashtags (ties for \code{n}th position will be included) were used in tweets.
}
\examples{
\dontrun{

top_n_hashtags(sqlite_con, n = 10)
my_plot <- top_n_hashtags(sqlite_con, 20)

}

}
