% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_liked_tweets.R
\name{top_n_liked_tweets}
\alias{top_n_liked_tweets}
\title{Plot top n liked tweets}
\usage{
top_n_liked_tweets(sqlite_con, n = 10, tweet_chars = 80, chars_per_line = 50)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}

\item{n}{Number of tweets to be plotted. Note, ties will be included.
Default value is 10.}

\item{tweet_chars}{Number of characters of the tweet text to be displayed.}

\item{chars_per_line}{How many characters of the tweet text to be displayed
per line without breaking a word.}
}
\value{
ggplot2 plot.
}
\description{
Create a ggplot2 bar chart of the top \code{n} liked tweets (ties for
\code{n}th position will be included), based on Twitter metrics.
}
\examples{
\dontrun{

top_n_liked_tweets(sqlite_con, 20)

}

}
