% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_replied_to_tweets.R
\name{top_n_replied_to_tweets}
\alias{top_n_replied_to_tweets}
\title{Plot top n replied to tweets}
\usage{
top_n_replied_to_tweets(
  sqlite_con,
  n = 10,
  tweet_chars = 80,
  chars_per_line = 50,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}

\item{n}{Number of tweets to be plotted. Note, ties will be included.
Default value is 10. Recommended value to ensure clear visualisation is
<= 10.}

\item{tweet_chars}{Number of characters of the tweet text to be displayed.}

\item{chars_per_line}{How many characters of the tweet text to be displayed
per line without breaking a word.}

\item{return_data}{Should the data underlying the chart be returned?
The default is \code{FALSE}. If \code{return_data = TRUE}, the data can be accessed
in the second element, \code{data}, of the returned list.}

\item{...}{Other arguments to be passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}.}
}
\value{
ggplot2 plot. If \code{return_data = TRUE}, returns a named list with the
first element, \code{chart}, being a ggplot2 plot, and the second element,
\code{data}, being the underlying data for the ggplot2 plot in the form of a
data frame.
}
\description{
Create a ggplot2 bar chart of the top \code{n} replied to tweets
(ties for \code{n}th position will be included), based on Twitter metrics.
}
\examples{
\dontrun{

top_n_replied_to_tweets(sqlite_con, 12)

top_n_replied_to_tweets(sqlite_con, fill = "blue")

top_10_replied_to_tweets <- top_n_replied_to_tweets(sqlite_con, return_data = TRUE)

}

}
