% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetexploR-package.R
\docType{package}
\name{tweetexploR-package}
\alias{tweetexploR}
\alias{tweetexploR-package}
\title{tweetexploR: Explore Tidy Tweet SQLite Databases}
\description{
Explore and visualise a collection of tweets that has been tidied into an SQLite database file by the [tidy-tweet package](https://github.com/QUT-Digital-Observatory/tidy_tweet).
}
\author{
\strong{Maintainer}: QUT Digital Observatory \email{digitalobservatory@qut.edu.au}

}
\keyword{internal}
