% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_workflow.R
\name{num_tweets_by_username}
\alias{num_tweets_by_username}
\title{Plot number of tweets for top n usernames}
\usage{
num_tweets_by_username(sqlite_con, n)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}

\item{n}{Number of usernames to be plotted. Note, ties will be included.}
}
\value{
ggplot2 plot.
}
\description{
Create a ggplot2 bar chart of the number of tweets per username
for the top \code{n} usernames (ties for \code{n}th position will be included).
}
\examples{
\dontrun{

num_tweets_by_username(sqlite_con, n = 10)
my_plot <- num_tweets_by_username(sqlite_con, 20)
}

}
