% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_mentions.R
\name{top_n_mentions}
\alias{top_n_mentions}
\title{Plot top n account mentions}
\usage{
top_n_mentions(sqlite_con, n = 10, return_data = FALSE, ...)
}
\arguments{
\item{sqlite_con}{\href{https://rsqlite.r-dbi.org/reference/sqliteconnection-class}{Class SQLiteConnection}
object that is a connection to an SQLite .db file created by the
\href{https://github.com/QUT-Digital-Observatory/tidy_tweet}{tidy-tweet package}.
The database contains a collection of tweets in relational tables. This can
be created with \code{\link[=connect_to_sqlite_db]{connect_to_sqlite_db()}}.}

\item{n}{Number of accounts to be plotted. Note, ties will be included.
Default value is 10.}

\item{return_data}{Should the data underlying the chart be returned?
The default is \code{FALSE}. If \code{return_data = TRUE}, the data can be accessed
in the second element, \code{data}, of the returned list.}

\item{...}{Other arguments passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}.}
}
\value{
ggplot2 plot. If \code{return_data = TRUE}, returns a named list with the
first element, \code{chart}, being a ggplot2 plot, and the second element,
\code{data}, being the underlying data for the ggplot2 plot in the form of a
data frame.
}
\description{
Create a ggplot2 bar chart of the number of times the top \code{n}
accounts (ties for \code{n}th position will be included) were mentioned in
tweets.
}
\examples{
\dontrun{

top_n_mentions(sqlite_con, n = 12)

my_plot <- top_n_mentions(sqlite_con, 20)

top_n_mentions(sqlite_con, fill = "blue")

top_10_mentions <- top_n_mentions(sqlitecon, return_data = TRUE)

}

}
